Roughly modularizes project
